Add ability to update asset data

+++ src/scenes/AddItem/ItemForm/index.js
@@ -27,7 +27,8 @@ import LinearGradient from 'react-native-linear-gradient';
 import FeatherIcon from 'react-native-vector-icons/Feather';
 
 import { isIphoneX } from '~/global/device';
-import { CREATE_ASSET } from '~/graphql/assets/mutations';
+import { GET_COMPANY_ASSETS } from '~/graphql/assets/queries';
+import { CREATE_ASSET, UPDATE_ASSET } from '~/graphql/assets/mutations';
 import { isSmallDevice, convertToApolloUpload } from '~/global/utils';
 import colors from '~/global/colors';
 import assets from '~/global/assets';
@@ -214,7 +215,7 @@ class ItemForm extends Component<Props, State> {
     const item = navigation.getParam('item', null);
 
     if (item) {
-      // console.log(item)
+      // console.log(item);
       const itemCopy = { ...item };
       const {
         name,
@@ -241,12 +242,10 @@ class ItemForm extends Component<Props, State> {
         navigation.setParams({ userCanDelete: true, userCanEdit: true });
       }
 
-      // console.log(photos);
+      // console.log(1, photos);
       // console.log(photosOfDamages);
-      itemCopy.photos = photos.map(({ url }) => ({ uri: `https://api.staging.inventoryapp.info${url}` })) || [];
-      itemCopy.photosOfDamages = photosOfDamages.map(({ url }) => ({
-        uri: `https://api.staging.inventoryapp.info${url}`,
-      })) || [];
+      itemCopy.photos = photos.map(url => ({ uri: url }));
+      itemCopy.photosOfDamages = photosOfDamages.map(url => ({ uri: url }));
       itemCopy.responsibleId = responsible;
       itemCopy.status = status === 'on_processing'
         ? constants.placeholders.status.onProcessing
@@ -345,8 +344,8 @@ class ItemForm extends Component<Props, State> {
      * in this.checkFields and this.checkForErrors
      */
 
-    const { createAsset } = this.props;
-    const { photos, photosOfDamages } = this.state;
+    const { createAsset, updateAsset } = this.props;
+    const { photos, photosOfDamages, id: assetId } = this.state;
 
     const isFormInvalid = await Promise.resolve()
       .then(() => this.checkFields())
@@ -358,8 +357,16 @@ class ItemForm extends Component<Props, State> {
           company: { id: companyId },
         },
       } = this.props;
-      // $FlowFixMe
-      const variables = pick(constants.createAssetNecessaryProperties, this.state);
+
+      let variables;
+
+      if (assetId) {
+        // $FlowFixMe
+        variables = pick(constants.createAssetNecessaryProperties.concat('id'), this.state);
+      } else {
+        // $FlowFixMe
+        variables = pick(constants.createAssetNecessaryProperties, this.state);
+      }
 
       variables.companyId = companyId;
       // $FlowFixMe
@@ -407,27 +414,29 @@ class ItemForm extends Component<Props, State> {
           : 'on_processing';
       }
 
-      if (photos.length) {
-        try {
-          const photosResult = await convertToApolloUpload(photos, '.');
-          variables.photos = photosResult;
-        } catch (error) {
-          console.log(error.message);
-        }
-      }
-
-      if (photosOfDamages.length) {
-        try {
-          const defectsResult = await convertToApolloUpload(photosOfDamages, '.');
-          variables.photosOfDamages = defectsResult;
-        } catch (error) {
-          console.log(error.message);
-        }
-      }
-
-      // console.log(variables);
       try {
-        await createAsset({ variables });
+        const photosResult = await convertToApolloUpload(photos, '.');
+        variables.photos = photosResult;
+      } catch (error) {
+        console.log(error.message);
+      }
+
+      try {
+        const defectsResult = await convertToApolloUpload(photosOfDamages, '.');
+        variables.photosOfDamages = defectsResult;
+      } catch (error) {
+        console.log(error.message);
+      }
+
+      console.log(variables);
+      try {
+        let response;
+        if (assetId) {
+          response = await updateAsset({ variables });
+        } else {
+          response = await createAsset({ variables });
+        }
+        console.log(response);
       } catch (error) {
         Alert.alert(error.message);
         console.dir(error);
@@ -549,6 +558,8 @@ class ItemForm extends Component<Props, State> {
       },
     } = this;
 
+    const isLocalFile = uri.startsWith('file');
+
     return (!uri && isNewItem) || (!uri && userRole === constants.roles.admin) ? (
       <TouchableOpacity style={styles.addPhotoBarButton} onPress={this.handleAddPhoto}>
         <IonIcon size={26} {...iconProps} name="ios-add-circle" color={colors.border} />
@@ -559,7 +570,7 @@ class ItemForm extends Component<Props, State> {
           <TouchableOpacity
             activeOpacity={0.5}
             style={[styles.removePhotoIcon, isSmallDevice && styles.smallerIcon]}
-            onPress={() => this.handleRemovePreviewPhotoBarItem(index)}
+            onPress={() => this.handleRemovePreviewPhotoBarItem(index, isLocalFile)}
           >
             <Image source={assets.deletePhoto} />
           </TouchableOpacity>
@@ -583,19 +594,21 @@ class ItemForm extends Component<Props, State> {
     );
   };
 
-  handleRemovePreviewPhotoBarItem = async (removedIndex: number) => {
+  handleRemovePreviewPhotoBarItem = async (removedIndex: number, isLocalFile: boolean) => {
     const {
-      state: { showPhotos, activePreviewIndex },
       state,
+      state: { showPhotos, activePreviewIndex },
     } = this;
     const currentlyActive = showPhotos ? 'photos' : 'photosOfDamages';
     // $FlowFixMe
     const { uri } = state[currentlyActive][removedIndex];
 
-    try {
-      RNFS.unlink(uri);
-    } catch (error) {
-      Alert.alert(error.message);
+    if (isLocalFile) {
+      try {
+        RNFS.unlink(uri);
+      } catch (error) {
+        Alert.alert(error.message);
+      }
     }
 
     this.setState(state => ({
@@ -807,7 +820,9 @@ class ItemForm extends Component<Props, State> {
             </View>
             {(showSaveButton || isNewItem) && (
               <TouchableOpacity style={styles.saveItem} onPress={this.handleSubmitForm}>
-                <Text style={styles.saveItemText}>{constants.buttonTitles.saveItem}</Text>
+                <Text style={styles.saveItemText}>
+                  {isNewItem ? constants.buttonTitles.saveItem : constants.buttonTitles.saveChanges}
+                </Text>
               </TouchableOpacity>
             )}
             <DateTimePicker
@@ -835,13 +850,10 @@ export default compose(
     // $FlowFixMe
     props: ({ data: { userCompany } }) => ({ userCompany }),
   }),
-  // graphql(QUERIES.GET_CURRENT_USER_COMPANIES, {
-  //   // $FlowFixMe
-  //   props: ({ data: { current } }) => ({ currentUser: current }),
-  // }),
   graphql(QUERIES.GET_USER_ID_CLIENT, {
     // $FlowFixMe
     props: ({ data: { id } }) => ({ currentUserId: id }),
   }),
   graphql(CREATE_ASSET, { name: 'createAsset' }),
+  graphql(UPDATE_ASSET, { name: 'updateAsset', refetchQueries: [GET_COMPANY_ASSETS] }),
 )(ItemForm);
